+++ ui/src/main/preload.js
@@ -36,6 +36,8 @@ contextBridge.exposeInMainWorld('bsr', {
 
   listLibrary: () => ipcRenderer.invoke('library:list'),
   openLibrary: (dir) => ipcRenderer.invoke('library:open', { dir }),
+  searchLibrary: (query, options) =>
+    ipcRenderer.invoke('library:search', { query, options }),
   renameSession: (name) => ipcRenderer.invoke('session:rename', { name }),
   undo: () => ipcRenderer.invoke('edit:undo'),
   undoDepth: () => ipcRenderer.invoke('edit:undoDepth'),